+++ JsonToolsNppPlugin/JSONTools/RandomJsonFromSchema.cs
@@ -19,7 +19,7 @@ namespace JSON_Tools.JSON_Tools
         /// <summary>
         /// 50% chance for true or false
         /// </summary>
-        private static JNode RandomBoolean(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomBoolean(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             if (random.NextDouble() < 0.5)
                 return new JNode(true, Dtype.BOOL, 0);
@@ -29,7 +29,7 @@ namespace JSON_Tools.JSON_Tools
         /// <summary>
         /// random double from -5 to 5
         /// </summary>
-        private static JNode RandomFloat(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomFloat(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             return new JNode(random.NextDouble() * 10 - 5, Dtype.FLOAT, 0);
         }
@@ -37,7 +37,7 @@ namespace JSON_Tools.JSON_Tools
         /// <summary>
         /// random int from -1 million to 1 million 
         /// </summary>
-        private static JNode RandomInt(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomInt(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             return new JNode(random.Next(-1_000_000, 1_000_001), Dtype.INT, 0);
         }
@@ -45,17 +45,17 @@ namespace JSON_Tools.JSON_Tools
         /// <summary>
         /// 50% chance of random int, 50% chance of random float
         /// </summary>
-        private static JNode RandomNumber(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomNumber(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             if (random.NextDouble() < 0.5)
-                return RandomInt(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-            return RandomFloat(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
+                return RandomInt(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+            return RandomFloat(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
         }
 
         /// <summary>
         /// always returns the null JNode
         /// </summary>
-        private static JNode RandomNull(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomNull(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             return new JNode();
         }
@@ -68,7 +68,7 @@ namespace JSON_Tools.JSON_Tools
         /// string of random length between 0 and 10<br></br>
         /// Can contain any printable ASCII character, possibly multiple times
         /// </summary>
-        private static JNode RandomString(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomString(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             int length = random.Next(11);
             StringBuilder sb = new StringBuilder(length);
@@ -93,7 +93,7 @@ namespace JSON_Tools.JSON_Tools
         #endregion
 
         #region RANDOM_ITERABLES
-        private static JNode RandomArray(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomArray(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             Dictionary<string, JNode> children = ((JObject)schema).children;
             if (!children.TryGetValue("items", out JNode items))
@@ -123,10 +123,10 @@ namespace JSON_Tools.JSON_Tools
                 int ncontains = random.Next(minContains, maxContains + 1);
                 length -= ncontains;
                 for (int ii = 0; ii < ncontains; ii++)
-                    outItems.Add(RandomJson(contains, minArrayLength, maxArrayLength, extended_ascii_strings));
+                    outItems.Add(RandomJsonHelper(contains, refs, minArrayLength, maxArrayLength, extended_ascii_strings));
             }
             for (int ii = 0; ii < length; ii++)
-                outItems.Add(RandomJson(items, minArrayLength, maxArrayLength, extended_ascii_strings));
+                outItems.Add(RandomJsonHelper(items, refs, minArrayLength, maxArrayLength, extended_ascii_strings));
             if (contains != null)
             {
                 // shuffle the array to randomize the order in which
@@ -142,7 +142,7 @@ namespace JSON_Tools.JSON_Tools
             return new JArray(0, outItems);
         }
 
-        private static JNode RandomObject(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomObject(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             var children = ((JObject)schema).children;
             if (!children.TryGetValue("properties", out JNode properties))
@@ -186,13 +186,13 @@ namespace JSON_Tools.JSON_Tools
             foreach (string k in requiredKeys.Concat(optionalIncluded))
             {
                 JNode subschema = propertiesObj[k];
-                result[k] = RandomJson(subschema, minArrayLength, maxArrayLength, extended_ascii_strings);
+                result[k] = RandomJsonHelper(subschema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
             }
             return new JObject(0, result);
         }
         #endregion
 
-        private static readonly Dictionary<string, Func<JNode, int, int, bool, JNode>> GENERATORS = new Dictionary<string, Func<JNode, int, int, bool, JNode>>
+        private static readonly Dictionary<string, Func<JNode, JObject, int, int, bool, JNode>> GENERATORS = new Dictionary<string, Func<JNode, JObject, int, int, bool, JNode>>
         {
             { "array", RandomArray },
             { "boolean", RandomBoolean },
@@ -211,18 +211,18 @@ namespace JSON_Tools.JSON_Tools
         /// <param name="maxArrayLength"></param>
         /// <returns></returns>
         /// <exception cref="ArgumentOutOfRangeException"></exception>
-        private static JNode RandomAnything(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomAnything(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             int choice = random.Next(7);
             switch (choice)
             {
-                case 0: return RandomBoolean(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 1: return RandomFloat(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 2: return RandomInt(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 3: return RandomNull(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 4: return RandomString(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 5: return RandomArray(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
-                case 6: return RandomObject(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 0: return RandomBoolean(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 1: return RandomFloat(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 2: return RandomInt(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 3: return RandomNull(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 4: return RandomString(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 5: return RandomArray(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                case 6: return RandomObject(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
                 default: throw new ArgumentOutOfRangeException();
             }
         }
@@ -231,14 +231,14 @@ namespace JSON_Tools.JSON_Tools
         /// choose a random schema from an anyOf list of schemas, and make random JSON based on that schema
         /// </summary>
         /// <returns></returns>
-        private static JNode RandomAnyOf(JNode anyOf, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        private static JNode RandomAnyOf(JNode anyOf, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             JArray anyOfArr = (JArray)anyOf;
             JNode schema = anyOfArr.children[random.Next(anyOfArr.Length)];
-            return RandomJson(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
+            return RandomJsonHelper(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
         }
 
-        public static JNode RandomJson(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        public static JNode RandomJsonHelper(JNode schema, JObject refs, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
         {
             if (!(schema is JObject obj))
             {
@@ -247,7 +247,7 @@ namespace JSON_Tools.JSON_Tools
                     throw new SchemaValidationException("A JSON schema must be an object or a boolean");
                 }
                 if ((bool)schema.value)
-                    return RandomAnything(new JObject(), minArrayLength, maxArrayLength, extended_ascii_strings);
+                    return RandomAnything(new JObject(), new JObject(), minArrayLength, maxArrayLength, extended_ascii_strings);
                     // the true schema validates everything, so we'll make a
                     // random instance of a random type
                 return new JNode();
@@ -255,12 +255,19 @@ namespace JSON_Tools.JSON_Tools
                 // and the null JNode is the closest we can get to nothing
             }
             if (obj.Length == 0) // the empty schema validates everything
-                return RandomAnything(new JObject(), minArrayLength, maxArrayLength, extended_ascii_strings);
+                return RandomAnything(new JObject(), new JObject(), minArrayLength, maxArrayLength, extended_ascii_strings);
             if (!obj.children.TryGetValue("type", out JNode typeNode))
             {
                 if (!obj.children.TryGetValue("anyOf", out JNode anyOf))
-                    throw new SchemaValidationException("A schema must have an \"anyOf\" keyword or a \"type\" keyword");
-                return RandomAnyOf(anyOf, minArrayLength, maxArrayLength, extended_ascii_strings);
+                {
+                    if (!obj.children.TryGetValue("$ref", out JNode refnode))
+                        throw new SchemaValidationException("A schema must have an \"anyOf\" keyword or a \"type\" keyword");
+                    var refname = ((string)refnode.value).Split('/').Last();
+                    if (!refs.children.TryGetValue(refname, out JNode reference))
+                        throw new SchemaValidationException($"Reference {refname} to an undefined schema");
+                    return RandomJsonHelper(reference, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+                }
+                return RandomAnyOf(anyOf, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
             }
             if (obj.children.TryGetValue("enum", out JNode enum_))
             {
@@ -272,10 +279,20 @@ namespace JSON_Tools.JSON_Tools
             {
                 // multiple scalar types possible
                 JNode typeChoice = typeArr.children[random.Next(typeArr.Length)];
-                return GENERATORS[(string)typeChoice.value](schema, minArrayLength, maxArrayLength, extended_ascii_strings);
+                return GENERATORS[(string)typeChoice.value](schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
             }
             var typeGenerator = GENERATORS[(string)typeNode.value];
-            return typeGenerator(schema, minArrayLength, maxArrayLength, extended_ascii_strings);
+            return typeGenerator(schema, refs, minArrayLength, maxArrayLength, extended_ascii_strings);
+        }
+
+        public static JNode RandomJson(JNode schema, int minArrayLength, int maxArrayLength, bool extended_ascii_strings)
+        {
+            var schemobj = (JObject)schema;
+            JNode refs;
+            if (!(schemobj.children.TryGetValue("$defs", out refs)
+                || schemobj.children.TryGetValue("definitions", out refs)))
+                refs = new JObject();
+            return RandomJsonHelper(schema, (JObject)refs, minArrayLength, maxArrayLength, extended_ascii_strings);
         }
     }
 }